Create Technical_and_Policy_Framework_for_IXP.md
The Internet Society has been helping address connectivity gaps for many years, including building and supporting Internet Exchange Points (IXPs) to improve traffic flow and help people get cheaper, faster, and better Internet service. We would like to showcase and invite the Peering community to guide and  collaborate with the following whitepaper: Technical and Operational Policy Framework for an IXP. This document outlines the technical specifications, connection policies, and operational guidelines for participants of the Internet Exchange Point (IXP). It aims to ensure efficient, secure, and fair interconnection among members. We will present the outline of the document and ask for feedback. We are doing this presentation in several regions to have a worldwide overview of what is useful (and what is not). The outcome of this effort will be a publicly available document.